Unify Rydberg and Transmon Hamiltonians to ``HardwareHamiltonian`` (#3930)
* baseline to allow-... pr

* renamed files

* typo in doc string

* pylint

* hard-copy and paste again for readable diffs on github

* hard-copy and paste again for readable diffs on github

* hard-copy and paste again for readable diffs on github

* double-copied line

* tests passing

* changelog

* typo

* typos

* terms in theo. backg. positive

* minus sign confusion

* NOW the signs are correct

* Update doc/releases/changelog-dev.md


* code review suggestion: one sum

* Apply suggestions from code review


* refactor rydberg in separate file

* separate rydberg test file

* internal class docstring

* minor

* black

* Update regex match in failing test

* Update doc/releases/changelog-dev.md


* Apply suggestions from code review


* update example code updates

* Update pennylane/pulse/hardware_hamiltonian.py


* Update pennylane/pulse/hardware_hamiltonian.py


---------
