+++ pennylane/pulse/__init__.py
@@ -68,7 +68,7 @@ Hardware Compatible Hamiltonians
     :toctree: api
 
     ~rydberg_interaction
-    ~rydberg_drive
+    ~drive
 
 
 Creating a parametrized Hamiltonian
@@ -259,4 +259,5 @@ JIT-compiling is optional, and one can remove the decorator when only single exe
 from .convenience_functions import constant, pwc, pwc_from_function, rect
 from .parametrized_evolution import ParametrizedEvolution
 from .parametrized_hamiltonian import ParametrizedHamiltonian
-from .rydberg_hamiltonian import RydbergHamiltonian, rydberg_drive, rydberg_interaction
+from .hardware_hamiltonian import HardwareHamiltonian, drive
+from .rydberg import rydberg_interaction
